Fix: VM cleanup and separation of task
